read_file/requirements.txt
PyMuPDF
python-docx
easyocr
pywin32
numpy
opencv-python
pillow